Add PoC of commit count graph
This proof of concepts shows the creation
of line graphs depicting the number of commits
of an author per active day.

Upon using the command line flag --img,
one graph exported to a .png file per author will be created.

Note that the addition adds the dependence to the following go package:
* github.com/wcharczuk/go-chart

Known limitations:
* If an author has only one single day of contributions, the graph won't display
* The y axis shows float numbers even though commit count is a discrete value
* It doesn't look very nice

Signed-off-by: David <david@isan.engineer>